+++ src/tools.js
@@ -2,5 +2,6 @@
 
 const { createBrowsingTools } = require('../tools/browse');
 const { createMobileTools } = require('../tools/mobile');
+const { createShellTools } = require('../tools/shell');
 
-module.exports = { createBrowsingTools, createMobileTools };
+module.exports = { createBrowsingTools, createMobileTools, createShellTools };